2D Depth Map
The piece of code where imread was not sucessful is fixed. A spagethi
code of getting the total value of a column, sum it and averaged it,
spot the value in a Mat has added. Basically is something similar to
horizontal Histogram like Laura's. Need to refactor. May add a
commandline for convenience.