new RawData
Added new set of Raw data. new set (settwo) is angled at 90degree,
distanced at 68.4cm. Some images are duplicates to have to labels, one
on the left and the other on the right, for stitching. To stitch, you
may need to combine those picture to have a single image with two
markers before stitching.