FFT NOT DONE
Im sorry I lied. Lots of things not done. ReMap wasn't used and was not
functioning. Result is still a mess. Can't reverse FFT to the original
image let alone FFT to the right domain. Will work on it tomorrow. On
the side note, "Window" correlation still takes too long. Need to work
on it. NO MORE TIME