image registration
Depth data is now resized to half the size of color resolution, while
color is downsampled to half its size. So called image registration. The
actual way of doing it is referring its UV value in UVMap and find the
corresponding value, but there's no such precise need as no point cloud
is being used here.